Add findDocumentSymbols
A question was asked in gitter if VS Code has an equivalent
of WebStorm's "File Structure Popup" in HTML. The closest that vscode has
is "Show document symbols", but the html language service does not
provide that.

To attempt parity with webstorm in this area, I have implemented
findDocumentSymbols. I have augmented the parser to track tag attributes
and their values (I could see nowhere before that attributes were being
used) and combine the tag name with id and class attributes to build
a symbol name.

For example, <div id='test' class='my classes'></div>
will have the symbol name div#id.my.classes. For child elements,
I set the containerName to be the symbol name of their parent tag.
So in <div id='container1'><span></span></div>, the span element
would have the symbol name span and the container name div#container1

I set the SymbolKind to field just to have something as I am unsure
what the convention is for non-code related symbols.

For this to appear in vscode, it would necessitate a change there as well.
I have a pull request ready to submit if this is accepted.