Allow formatting to end inside an element
Previously, if a range started and ended inside an element it would not be formatted.
However, if it started inside an element but ended outside it would format.

With this change as long as the range starts outside on element it will format.
This is safe as the beautifier formats incomplete html just fine up to where it ends.